store/wizard: move output selectors to dedicated module
The output selectors were defined in the monolithic wizard slice file
alongside unrelated selectors. Co-locating them with the rest of the
output module keeps things consistent and easier to find.
